+++ packages/msw/unit.d.ts
@@ -1,0 +1,5 @@
+import { SetupServerApi } from 'msw/node';
+
+declare module 'hops-msw/unit' {
+  export const getMockServer: () => SetupServerApi;
+}